moved functions to a new file
